Updated python scripts to work with protobuf 3.2.0
Since the build script of python grpcio-tools package
is very much entangled with the build system of protobuf,
and protobuf build scripts added a build step that
compiles js/embed.cc to create a preprocessor binary
that takes some .js files and generates a single
well_known_types_embed.cc which then gets linked to the
rest of protobuf source tree.

The generated code seems to be used only by the JS
portion of protoc, so is not expected to affect
our python build.  Since we did not want to replicate
the entire build process of protobuf, we decided to
pre-generate the well_known_types_embed.cc and keep
it in our repo and then substitute the js/embed.cc
compiler file with it.