Use grpc_1_0 flag in beta_python_plugin_test
Beta code elements are not generated at all in _pb2_grpc.py files.

This duplicates a lot of the in-test code generation done in
_split_definitions_test. In a future clean-up we may want to
deduplicate the common behavior, put it in a module available to all
other tests, and do all of our testing of generated code with in-test
code generation.