Ignore zlib submodule if its dirty (due to CMake)
When using CMake to build, the zlib submodule ends up with a generated
file that makes Git consider the submodule dirty. This state can be
ignored for day-to-day development on gRPC.

We cannot use the weaker "ignore = untracked" as the build process
deletes a tracked file as well as generating a new file.