Reorder fields in slice to share the same bytes for length fields.
Ref-counted and inlined slices both have a lenght value, but
they are not in the same byte offset. As a result, we will
have two different loads based on a branch.

Instead move them to the same byteoffset, so that we will have
one move and the same number of branches.

Difference can be seen on:
https://godbolt.org/z/kqFZcP