Ruby tooling: respect user toolchain overrides
While compilation flag overrides for the Ruby native extension are
currently functional, specifying an alternate compiler is not, as the
Rbconfig values for key toolchain binaries are hardcoded at compile time
of the Ruby interpreter.  This patch allows them to be overrridden on
the command line via standard environment variables, defaulting to the
Rbconfig values only if unspecified by the user.