Add UserState dictionary to C# ServerCallContext
This commit adds a IDictionary<object, object> UserState member to the
ServerCallContext. Interceptors and call handlers can use this member to
pass per-call state between themselves.

Like other members of ServerCallContext, UserState is not thread-safe.

UserState is initialized on demand so that calls that don't use
UserState don't need to pay for it.

Closes https://github.com/grpc/grpc/issues/17759